requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
cycler==0.10.0
demjson==2.2.4
kiwisolver==1.1.0
matplotlib==3.1.1
nose==1.3.7
numpy==1.17.0
opencv-contrib-python==4.1.0.25
pandas==0.25.1
Pillow==8.2.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
toml==0.10.0
virtualenv==16.7.4
xlrd==1.2.0
xmltodict==0.12.0
yagmail==0.11.220
yapf==0.28.0
mysql.connector==